Refactored TreeTypeMap
Goes into a separate source files. Several simplifying refactorings.
